--- styles/download.css
@@ -1,10 +1,0 @@
-.download-container {
-  position: absolute;
-  top: -64px;
-  left: 54px;
-}
-
-.download-icon {
-  width: 54px;
-  cursor: pointer;
-}
